Extract policies for memberships to module
