+++ app/policies/approves_policy.rb
@@ -1,4 +1,5 @@
 class ApprovesPolicy < ApplicationPolicy
+  include MembershipApproval
   attr_accessor :user, :membership
 
   def initialize(user, membership)
@@ -8,26 +9,5 @@ class ApprovesPolicy < ApplicationPolicy
 
   def update?
     all_user_validations
-  end
-
-  private
-
-  def all_user_validations
-    user_is_involved? &&
-      approver_and_requestor_are_not_both_admins? &&
-      user_is_not_requestor?
-  end
-
-  def user_is_involved?
-    !user.nil? && membership.users.include?(user)
-  end
-
-  def approver_and_requestor_are_not_both_admins?
-    admins = membership.league.admins
-    !(admins.include?(membership.requestor) && admins.include?(user))
-  end
-
-  def user_is_not_requestor?
-    membership.requestor != user
   end
 end
